requirements.txt
aiohttp==3.12.15
python-dotenv==1.1.1
websockets==15.0.1
numpy>=1.26.0
tabulate>=0.9.0